Add rollup map to PagedResults response
* Added value rollups to delegation/redelegation/unbonding APIS for both Accounts and Validators
  * Generic map to add different values for different PagedResults responses as needed
* General cleanup of APIs
* Updated token stats to use correct Bonded value
* Updated params to represent percentage values better
* Updated Migration 1.41 to include exception handling in procedures
* Updated BlockRetry to use the main block insertion method rather than the intermediary

closes: #290

+++ service/src/main/kotlin/io/provenance/explorer/domain/models/explorer/ExplorerModels.kt
@@ -34,8 +34,8 @@ data class Spotlight(
 
 data class GasStatistics(
     val time: String,
-    val minGasPrice: Int,
-    val maxGasPrice: Int,
+    val minGasPrice: BigDecimal,
+    val maxGasPrice: BigDecimal,
     val averageGasPrice: BigDecimal
 )
